Move tagsources window into mainwin
Moved releasewidget as well (which is on the tagsources window.
Refactored all references to tagsources and releasewidget.
TODO: Rename WEBDB Translation group if needed,
